Moved most of the project to /src/
